Kreiran kontroler i API rute za Product

+++ digitalProductsOnlineStore_2021_0179/app/Http/Controllers/ProductController.php
@@ -11,36 +11,62 @@ use App\Http\Resources\ProductCollection;
 class ProductController extends Controller
 {
     public function index(){
+
         $products = Product::all();
         return new ProductCollection($products);  
+
     }
     public function store(Request $request){
         $validated = $request->validate([
             'name' => 'required|string|max:255',
             'description' => 'required|string',
             'price' => 'required|numeric|min:0',
-            'category_id' => 'required|exists:categories_id',
+            'category_id' => 'required|exists:categories,id',
         ]);
         $product = Product::create($validated);
         return  new ProductResource($product);
     }
-    public function show(Product $product){
-        // $product = Product::find($id);
-        return new ProductResource($product);
+    public function show($id){
+        
+            $product = Product::find($id);
+        
+            if (!$product) {
+                return response()->json(['message' => 'Product not found'], 404);
+            }
+        
+         return new ProductResource($product);
+        
     }
-    public function update(Request $request, Product $product){
+    public function update(Request $request,  $id){
+        $product = Product::find($id);
+
+        if(!$product){
+            return response()->json(['message'=>'Product not found'], 404);
+        }
         $validated = $request->validate([
             'name' => 'string|max:255',
             'description' => 'string',
             'price' => 'numeric|min:0',
-            'category_id' => 'exists:categories_id',
+            'category_id' => 'exists:categories,id',
         ]);
+
         $product->update($validated);
-        return new ProductResource($product);
+        return response()->json(new ProductResource($product->fresh()), 200);
     }
-    public function destroy(Product $product){
+
+
+    public function destroy($id){
+    
+        $product = Product::find($id);
+
+         if (!$product) {
+             return response()->json(['message' => 'Product not found'], 404);
+         }
+
         $product->delete();
-        return response()->json(['message'=>'Product deleted succesfully', Response::HTTP_NO_CONTENT]);
+
+         return response()->json(['message' => 'Product deleted successfully'], 200);
+
         
     }
 }
